Osa4, tehtävä 4.21

+++ Osa4/controllers/blogs.js
@@ -39,8 +39,19 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({error: 'token missing or invalid'})
+  }
+  const blogToBeDeleted = await Blog.findById(request.params.id)
+  const user = await User.findById(decodedToken.id)
+  if (blogToBeDeleted.user.toString() !== user._id.toString()){
+    return response.status(401).json({error: 'unauthorized attempt to delete blog'})
+  }
+  else {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
